requirements.txt
# Core app dependencies
streamlit>=1.50.0
requests>=2.31.0
feedparser>=6.0.10
openai>=1.55.0
sentence-transformers>=3.0.0
pandas>=2.0.0
google-genai>=0.3.0
groq>=0.9.0
faiss-cpu>=1.7.4
bm25s>=0.2.0
tenacity>=8.2.3
python-dotenv>=1.0.0
boto3>=1.26.0
scikit-learn>=1.3.0
numpy>=1.24.0